requirements.txt
python-docx==0.8.11
requests==2.25.1
urllib3==1.26.4